added dashboard

+++ productsHome/models.py
@@ -5,7 +5,7 @@ from django.contrib.postgres.fields import ArrayField
 class Products(models.Model):
     title = models.CharField(max_length = 100)
     description = models.TextField()
-    views = models.IntegerField(null = True)
+    views = models.FloatField(null = True)
     image = models.ImageField(upload_to = 'productImages/')
     dateUploaded =  models.DateField(auto_now_add=True)
     Daytimestamp = models.DateField(auto_now_add=True)
@@ -16,12 +16,6 @@ class Products(models.Model):
 
 class Views(models.Model):
     X = ArrayField(models.IntegerField(blank = True))
-    # Y = ArrayField(models.DecimalField(blank = True,max_digits=2, decimal_places=2))
     Y = ArrayField(models.FloatField(blank= True))
     Viewtimestamp = models.DateTimeField(auto_now_add=True)
     
-
-# class Dashboard(models.Model):
-#     videoNumber = models.ForeignKey(Products, on_delete=models.CASCADE)
-#     X = models.IntegerField(null = True)
-#     Y = models.IntegerField(null = True)
